migration vers TypeScript

+++ src/factories/InventoryEntryFactory.ts
@@ -5,8 +5,22 @@ import InventorySession from '../entities/InventorySession';
 import moment from 'moment';
 import InventoryEntry from '../entities/InventoryEntry';
 
+interface InventoryEntryTableDefinition {
+    id?: number;
+    inventory_id?: number;
+    article_barcode?: string;
+    article_name?: string;
+    article_image?: string;
+    article_unit?: number;
+    article_price?: number;
+    scanned_at?: string;
+    article_quantity?: number;
+    saved_at?: string;
+}
+
 export default class InventoryEntryFactory {
-    static instance = null;
+    private static instance: InventoryEntryFactory;
+    private db: Database;
 
     static sharedInstance() {
         if (InventoryEntryFactory.instance === null) {
@@ -19,8 +33,8 @@ export default class InventoryEntryFactory {
         this.db = Database.sharedInstance();
     }
 
-    async findByInventorySessionIdAndBarcode(inventorySessionId, barcode) {
-        const response = await this.db.executeQuery('SELECT * FROM `inventories_entries` WHERE `inventory_id` = ? AND `article_barcode` = ? ORDER BY scanned_at DESC', [inventorySessionId, barcode]);
+    async findByInventorySessionIdAndBarcode(inventorySessionId: number, barcode: string) {
+        const response = await this.db.executeQuery('SELECT * FROM `inventories_entries` WHERE `inventory_id` = ? AND `article_barcode` = ? ORDER BY scanned_at DESC', [inventorySessionId.toString(), barcode]);
         const entries = [];
         for (let i = 0; i < response[0].rows.length; i++) {
             entries.push(this._rowToObject(response[0].rows.item(i)));
@@ -29,12 +43,15 @@ export default class InventoryEntryFactory {
         return entries;
     }
 
-    async findForInventorySession(inventorySession) {
+    async findForInventorySession(inventorySession: InventorySession) {
+        if (!inventorySession.id) {
+            throw new Error("InventorySession has no id");
+        }
         return await this.findForInventorySessionId(inventorySession.id);
     }
 
-    async findForInventorySessionId(inventorySessionId) {
-        const response = await this.db.executeQuery('SELECT * FROM `inventories_entries` WHERE `inventory_id` = ? ORDER BY `saved_at` DESC', [inventorySessionId]);
+    async findForInventorySessionId(inventorySessionId: number) {
+        const response = await this.db.executeQuery('SELECT * FROM `inventories_entries` WHERE `inventory_id` = ? ORDER BY `saved_at` DESC', [inventorySessionId.toString()]);
         const entries = [];
         for (let i = 0; i < response[0].rows.length; i++) {
             entries.push(this._rowToObject(response[0].rows.item(i)));
@@ -53,7 +70,7 @@ export default class InventoryEntryFactory {
         return inventorySessions;
     }
 
-    async persist(object) {
+    async persist(object: InventoryEntry) {
         const params = this._objectToParams(object);
 
         const response = await this.db.executeQuery(
@@ -64,11 +81,11 @@ export default class InventoryEntryFactory {
         return;
     }
 
-    async update(object) {
+    async update(object: InventoryEntry) {
         const params = this._objectToParams(object);
         console.log(params);
 
-        const response = await this.db.executeQuery(
+        await this.db.executeQuery(
             'INSERT INTO `inventories_entries` VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)',
             params
         );
@@ -76,8 +93,11 @@ export default class InventoryEntryFactory {
         return;
     }
 
-    async delete(object) {
-        const params = [object.id];
+    async delete(object: InventoryEntry) {
+        if (!object.id) {
+            throw new Error("Cannot delete InventoryEntry with no id!");
+        }
+        const params: (string|number)[] = [object.id];
 
         const response = await this.db.executeQuery(
             'DELETE FROM `inventories_entries` WHERE `id` = ?;',
@@ -87,7 +107,7 @@ export default class InventoryEntryFactory {
         return;
     }
 
-    _rowToObject(row) {
+    _rowToObject(row: InventoryEntryTableDefinition) {
         console.debug("inventories_entries row", row);
         const entry = new InventoryEntry();
         entry.id = Number(row.id);
@@ -104,7 +124,7 @@ export default class InventoryEntryFactory {
         return entry;
     }
 
-    _objectToParams(object) {
+    _objectToParams(object: InventoryEntry): any[] {
         const row = {
             id: object.id,
             inventory_id: object.inventoryId,
@@ -113,9 +133,9 @@ export default class InventoryEntryFactory {
             article_image: null,
             article_unit: object.articleUnit,
             article_price: object.articlePrice,
-            scanned_at: object.scannedAt.format(Database.DATETIME_FORMAT),
+            scanned_at: object.scannedAt ? object.scannedAt.format(Database.DATETIME_FORMAT) : null,
             article_quantity: object.articleQuantity,
-            saved_at: object.savedAt.format(Database.DATETIME_FORMAT)
+            saved_at: object.savedAt ? object.savedAt.format(Database.DATETIME_FORMAT) : null
         };
 
         const params = Object.values(row);
